+++ backend/src/main/java/com/c24_39_t_webapp/restaurants/dtos/response/OrderDetailsResponseDto.java
@@ -8,5 +8,5 @@ public record OrderDetailsResponseDto(
         String productName,
         Integer quantity,
         BigDecimal productPrice,
-        Double subtotal
+        BigDecimal subtotal
     ) {}
